requirements.txt
# requirements.txt
tensorflow
pandas
numpy
opencv-python-headless
scikit-learn
streamlit
matplotlib
pillow